feat(driver): Refactor driver card list on client
- Removed driver list component to driver entity's widget segment
