+++ packages/client/src/widgets/driver/ui/index.ts
@@ -1,0 +1,1 @@
+export { DriverList } from './list';
